buny::imports::std

Function read

pub fn read<T: DeserializeOwned>(rid: i32) -> Result<T, BunyError>

Reads an object from a descriptor.

This function is exposed for the functions that the register_source macro generates and should not be used directly.