buny

Trait Source

pub trait Source {
    // Required methods
    fn new() -> Self;
    fn get_search_novel_list(
        &self,
        query: Option<String>,
        page: i32,
        filters: Vec<FilterValue>,
    ) -> Result<NovelPageResult>;
    fn get_novel_update(
        &self,
        novel: Novel,
        needs_details: bool,
        needs_chapters: bool,
        page: i32,
    ) -> Result<Novel>;
    fn get_chapter_content_list(
        &self,
        novel: Novel,
        chapter: Chapter,
    ) -> Result<Vec<ContentBlock>>;
}

The required functions an Buny source must implement.

Required Methods

fn new() -> Self

Called to initialize a source.

If a source requires any setup before other functions are called, it should happen here.

fn get_search_novel_list( &self, query: Option<String>, page: i32, filters: Vec<FilterValue>, ) -> Result<NovelPageResult>

Returns the novel for a search query with filters.

fn get_novel_update( &self, novel: Novel, needs_details: bool, needs_chapters: bool, page: i32, ) -> Result<Novel>

Updates a given novel with new details and chapters, as requested. Additiohally, the page parameter indicates which page of chapters to fetch if pagination is supported.

fn get_chapter_content_list( &self, novel: Novel, chapter: Chapter, ) -> Result<Vec<ContentBlock>>

Returns the content array for a given novel chapter.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.